requirements.txt
python-chess>=1.9
librosa>=0.10
moderngl>=5.8
numpy>=1.24
Pillow>=10.0
soundfile>=0.12
